removed unnecessary cast
git-svn-id: https://zmpp.svn.sourceforge.net/svnroot/zmpp/trunk@531 7df28f57-6730-0410-8a28-a3245024edd1

+++ zmpp/src/java/org/zmpp/vm/ClassicObjectTree.java
@@ -152,8 +152,7 @@ public class ClassicObjectTree extends AbstractObjectTree {
    */
   protected int getPropertyNum(final int propertyAddress) {    
     final int sizeByte = getMemory().readUnsignedByte(propertyAddress);
-    return (short) (sizeByte - 32 *
-    		(getPropertyLength(propertyAddress + 1) - 1));
+    return sizeByte - 32 * (getPropertyLength(propertyAddress + 1) - 1);
   }
   
   /**
